backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
openai>=1.12.0
tiktoken>=0.5.0
cohere>=4.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
